services/account-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic[email]==2.7.0
aioodbc==0.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
opencensus-ext-azure==1.1.13
opencensus-ext-fastapi==0.7.2
pyodbc==5.1.0
python-dotenv==1.0.1